Renderable trait.
