+++ src/core/buffer.rs
@@ -279,4 +279,30 @@ mod tests {
                 .unwrap();
         expect_that!(&buffer.len(), eq(28));
     }
+
+    #[derive(Debug, PartialEq, Clone, Copy)]
+    struct MyVertex {
+        _pos: [f32; 3],
+        _color: [f32; 4],
+    }
+
+    #[test]
+    fn immutable_buffer_from_struct_data() {
+        let instance = Instance::create().unwrap();
+        let buffer = ImmutableBuffer::from_data(
+            &instance,
+            &[
+                MyVertex {
+                    _pos: [1., 2., 3.],
+                    _color: [4., 5., 6., 7.],
+                },
+                MyVertex {
+                    _pos: [-1., 2., -3.],
+                    _color: [4., 5., 6., 7.],
+                },
+            ],
+        )
+        .unwrap();
+        expect_that!(&buffer.len(), eq(56));
+    }
 }
