+++ src/core/mod.rs
@@ -18,8 +18,5 @@ pub use texture_format::TextureFormat;
 mod buffer;
 pub use buffer::{BufferCreationError, BufferLength, ImmutableBuffer};
 
-mod mesh;
-pub use mesh::{Mesh, VertexCount};
-
 mod pipeline;
-pub use pipeline::{Pipeline, PipelineConfig, PipelineCreationError, ShaderConfig};
+pub use pipeline::{Pipeline, PipelineConfig, PipelineCreationError, Renderable, ShaderConfig};
